Updated code to work with new versions; new readme

+++ bayes_gmm/gaussian_components_diag.py
@@ -91,11 +91,11 @@ class GaussianComponentsDiag(object):
         assert len(prior.S_0.shape) == 1, "For diagonal covariance, S_0 needs to be vector."
 
         # Initialize attributes
-        self.m_N_numerators = np.zeros((self.K_max, self.D), np.float)
-        self.S_N_partials = np.zeros((self.K_max, self.D), np.float)
-        self.log_prod_vars = np.zeros(self.K_max, np.float)
-        self.inv_vars = np.zeros((self.K_max, self.D), np.float)
-        self.counts = np.zeros(self.K_max, np.int)
+        self.m_N_numerators = np.zeros((self.K_max, self.D), float)
+        self.S_N_partials = np.zeros((self.K_max, self.D), float)
+        self.log_prod_vars = np.zeros(self.K_max, float)
+        self.inv_vars = np.zeros((self.K_max, self.D), float)
+        self.counts = np.zeros(self.K_max, int)
 
         # Perform caching
         self._cache()
@@ -103,11 +103,11 @@ class GaussianComponentsDiag(object):
         # Initialize components based on `assignments`
         self.K = 0
         if assignments is None:
-            self.assignments = -1*np.ones(self.N, np.int)
+            self.assignments = -1*np.ones(self.N, int)
         else:
 
             # Check that assignments are valid
-            assignments = np.asarray(assignments, np.int)
+            assignments = np.asarray(assignments, int)
             assert (self.N, ) == assignments.shape
             # Apart from unassigned (-1), components should be labelled from 0
             assert set(assignments).difference([-1]) == set(range(assignments.max() + 1))
@@ -121,7 +121,7 @@ class GaussianComponentsDiag(object):
     def _cache(self):
         self._cached_prior_square_m_0 = np.square(self.prior.m_0)
 
-        self._cached_square = np.zeros((self.N, self.D), np.float)
+        self._cached_square = np.zeros((self.N, self.D), float)
         self._cached_square = np.square(self.X)
 
         n = np.concatenate([[1], np.arange(1, self.prior.v_0 + self.N + 2)])  # first element dud for indexing
@@ -129,7 +129,7 @@ class GaussianComponentsDiag(object):
         self._cached_gammaln_by_2 = gammaln(n/2.)
         self._cached_log_pi = math.log(np.pi)
 
-        self.cached_log_prior = np.zeros(self.N, np.float)
+        self.cached_log_prior = np.zeros(self.N, float)
         for i in range(self.N):
             self.cached_log_prior[i] = self.log_prior(i)
 
@@ -378,7 +378,7 @@ def log_post_pred_unvectorized(gmm, i):
     Return the same value as `GaussianComponentsDiag.log_post_pred` but using
     an unvectorized procedure, for testing purposes.
     """
-    post_pred = np.zeros(gmm.K, np.float)
+    post_pred = np.zeros(gmm.K, float)
     for k in range(gmm.K):
         post_pred[k] = gmm.log_post_pred_k(i, k)
     return post_pred
